API: Updated Api controllers

+++ app/controllers/api/v1/comments_controller.rb
@@ -6,17 +6,20 @@ class Api::V1::CommentsController < Api::ApiController
 
   def index
     @comments = Comment.all
+    authorize @comments, class_name: CommentPolicy
     success_response(@comments)
   end
 
   def show
+    authorize @comment, class_name: CommentPolicy
     success_response(@comments)
   end
 
   def create
     @comment = @project.comments.new(comment_params)
     @comment.commenter = current_user
-    
+    authorize @comment, class_name: CommentPolicy
+
     if @comment.save
       success_response(@comment, :created)
     else
@@ -25,6 +28,8 @@ class Api::V1::CommentsController < Api::ApiController
   end
 
   def update
+    authorize @comment, class_name: CommentPolicy
+
     if @comment.update(comment_params)
       success_response(@comment, :updated)
     else
@@ -33,6 +38,8 @@ class Api::V1::CommentsController < Api::ApiController
   end
 
   def destroy
+    authorize @comment, class_name: CommentPolicy
+
     if @comment.destroy
       success_response(@comment, :deleted)
     else
